+++ extension/storage/filestorage/factory.go
@@ -25,6 +25,15 @@ import (
 // The value of extension "type" in configuration.
 const typeStr config.Type = "file_storage"
 
+const (
+	// use default bbolt value
+	// https://github.com/etcd-io/bbolt/blob/d5db64bdbfdee1cb410894605f42ffef898f395d/cmd/bbolt/main.go#L1955
+	defaultMaxTransactionSize  = 65536
+	defaultReboundSizeBelowMib = 10
+	defaultReboundSizeAboveMib = 100
+	defaultCompactionInterval  = time.Second * 5
+)
+
 // NewFactory creates a factory for HostObserver extension.
 func NewFactory() component.ExtensionFactory {
 	return component.NewExtensionFactory(
@@ -38,14 +47,13 @@ func createDefaultConfig() config.Extension {
 		ExtensionSettings: config.NewExtensionSettings(config.NewComponentID(typeStr)),
 		Directory:         getDefaultDirectory(),
 		Compaction: &CompactionConfig{
-			Directory: getDefaultDirectory(),
-			OnStart:   false,
-			OnRebound: false,
-			// use default bbolt value
-			// https://github.com/etcd-io/bbolt/blob/d5db64bdbfdee1cb410894605f42ffef898f395d/cmd/bbolt/main.go#L1955
-			MaxTransactionSize:       65536,
-			ReboundSizeBelowMiB:      10,
-			ReboundTotalSizeAboveMiB: 100,
+			Directory:                getDefaultDirectory(),
+			OnStart:                  false,
+			OnRebound:                false,
+			MaxTransactionSize:       defaultMaxTransactionSize,
+			ReboundSizeBelowMiB:      defaultReboundSizeBelowMib,
+			ReboundTotalSizeAboveMiB: defaultReboundSizeAboveMib,
+			CheckInterval:            defaultCompactionInterval,
 		},
 		Timeout: time.Second,
 	}